Update tweet_historical_events.py
The previous code was checking only if event occurred in the exact date of today (2020-10-05), this slices the date and check if there is a month-day match (10-05) and consequently tweets the event. 

There could also be a problem on how the links have been written, I'll look into it and let you know.