Updated Tweet.py to post image instead of tweet
The program uploads an image instead of the tweet. Another parameter could be passed to upload image WITH a tweet.
It assumes that the generated images have the naming convention "post_{no}.png" and are all located in the same directory as the tweet.py file.
It also assumes that there exists atleast 1 image with the above mentioned naming convention in the directory.